refactor: change the sub_series inputs
The offset input is changed to a "shifted offset" which is a signed
integer. The shifted offset is used to shift the series by the offset
value. Also, if the series input is a sub_series, then the shifted
offset is just added to the offset of the sub_series, which make the
sub_series not lose the offset information of the parent series. Check
the test_series.cpp on the `SeriesTest::SubSeriesOfSubSeries` test case
for more information.