feat: open and save state file
The state file is a JSON file that contains the current state of the
application. The state includes the current list of assets, the list of
strategies, and the list of backtests which just includes the strategy
and the asset used for the backtest. When the file is opened, the
application will recalculate the backtesting results.