feat: string as a method in JSON config will also expand to a method
The string value in the method JSON config previously will became as a
key to find the method in the named_config_methods. Now, if the method
is not found in the named_config_methods, then the value will expand to
a method without parameters. The later just work with methods that have
default value for all parameters.