feat: backtests list setup can be loaded from file
The setup file is a CSV file that contains the list of backtests to be
loaded. The CSV must have Asset, and Strategy columns, and can
optionally include a Name column. The Asset is the name of the asset,
and the Strategy is the name of the strategy to be used for the creating
the backtest. The Name is optional and can be used to give a custom name
to the backtest.